requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
pytest>=7.0.0
pypdf>=3.0.0


python-dotenv>=1.0.0
schedule>=1.1.0
sendgrid>=6.10.0